requirements.txt
fastapi 
uvicorn 
sqlmodel 
bcrypt==4.0.1
Jinja2 
python-multipart 
aiofiles
itsdangerous
python-dotenv 
pymysql 
boto3 
google-cloud-vision 
ultralytics 
passlib